client-python/src/requirements.txt
click==8.1.2
Flask==2.1.1
grpcio==1.44.0
grpcio-tools==1.44.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
protobuf==3.20.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
